+++ tket/src/Transformations/ThreeQubitSquash.cpp
@@ -24,6 +24,7 @@
 #include <string>
 #include <tkassert/Assert.hpp>
 
+#include "BasicOptimisation.hpp"
 #include "Circuit/CircUtils.hpp"
 #include "Circuit/Circuit.hpp"
 #include "Circuit/DAGDefs.hpp"
@@ -119,6 +120,7 @@ static Circuit candidate_sub(const Circuit &circ) {
     // TODO: for now we decompose all the way to CX. In the future, it's worth
     // considering keeping TK2, and decomposing to CX (or other gates) later
     // when necessary.
+    normalise_TK2().apply(repl);
     decompose_TK2().apply(repl);
     clifford_simp(false).apply(repl);
     return repl;